Add guide on how to ssh into the TCM environment (#2)
We have some existing notes for accessing the SSH environment. But for those new to doing this, or who need to access multiple environments (not just TCM) they are a little incomplete.

This guide should hopefully get those new to doing this up and rolling. For those who know what they are doing it should also cover the details they need.