Add the document to the project
This adds a document detailing the release process for the SROC team.

It aims to cover both services the team manages. But for reference it has been written in the context of a pending Charging Module API release. It is likely to need alteration and correction when we come to run our first TCM release.