Add release process document (#12)
This change adds a document that details the release process for the [charging-module-api](https://github.com/DEFRA/charging-module-api) and the [tactical-charging-module](https://github.com/DEFRA/sroc-tcm-admin).

It's primarily for use by the development team as it details actions they need to take. However, it should be of use to everyone on the team.