Add guide on using Postman
Our team uses Postman to interact with the [Charging Module API](https://github.com/DEFRA/charging-module-api) when working on features, bugs etc. The current team when it first took over the project were handed some files. This gave us a way to interact with each endpoint, but authentication was a manual procress. Also they didn't always demonstrate what was a parameter in the paths, or what additional query params could be used.

By using the OpenAPI spec we generated and a pre-request script we have written, we feel we now have a way that all users can

- get up and running quickly
- see more of the functionality the API supports
- get a clearer idea of what is parametised
- avoid the need to manually copy bearer tokens into requests into requests
- maintain our postman setup in central place under source control

This change adds a guide for new members on how to setup Postman to work with the Charging Module API.