Add new database folder test query to project
We were recently asked by QA & Test to extract data on transactions so they could replicate real world tests of large billing runs. We also now have an understanding of how to connect to our databases in the non-production environments.

So this change creates a place to store that information and as the first item adds the script we used to extract the transaction data (connecting to the databases will come later!)