Move admin endpoints to admin path
Move the regime and authorised systems endpoints under the admin path. They are not for use by clients so there is not the need to manage their versioning in the same way.

This makes it easier on the development team and also clearer which parts of the API users can expect access to.