Remove unused endpoints from OpenAPI spec
https://trello.com/c/letq5mhW
https://github.com/DEFRA/charging-module-api/pull/153

https://trello.com/c/letq5mhW

We had developed some endpoints in the expectation they would be needed by our user. However, after a recent consultation, we have confirmed there are some that are not used or unlikely to be used for a very long time.

So the team have agreed that rather than continue to develop and maintain them we should remove them until they are actually needed.

This will reduce the complexity of the solution, and the maintenance burden on the tech team (development and testing).

This change updates the OpenAPI spec to match the changes made to the Charging Module API.