Underscores in urls? Oh my! (#27)
See [Underscores in urls? Oh my!](https://github.com/DEFRA/charging-module-api/pull/164)

Most likely it was because that's what the underlying controller/endpoints are called, we seem to have some endpoints that expect an underscore in the path.

It's perfectly permissible. However, it's definitely not the convention for URLs. To quote Google

> Consider using punctuation in your URLs. The URL **http://www.example.com/green-dress.html** is much more useful to us than **http://www.example.com/greendress.html**. We recommend that you use hyphens (-) instead of underscores (_) in your URLs.

So before the API hits a live production environment, we're taking the opportunity to update it to follow the current convention of using dashes.

There is also the 'blackboard' effect on us as a delivery team. It's for this same reason we added a [`PATCH` version of the bill run `/send` endpoint](https://github.com/DEFRA/charging-module-api/pull/156). 🧑‍🏫 😁