Start maintaining a draft version of OpenAPI spec
https://trello.com/c/cfA9RqKz

So the rest of the team and our stakeholders have visibility of what changes are coming in the next release of the Charging Module API we have agreed to start maintaining a `draft` version. As we make changes we'll re-generate the draft version and then update [SwaggerHub](https://app.swaggerhub.com/apis-docs/sro/charging-module_api/draft) with it.

This change covers generating the initial `draft` version and updating the OpenAPI README with details about it and what team members need to do.