Update main spec file version and description
We were intending to add instructions about updating the version and description each time we generate the `draft` version of the spec. But on reflection we decided it would be better to just have them in always, and add instructions to remove them when generating a release version.

We believe doing it this way round will make maintaining the spec easier.