Start adding `required` details to request schemas
https://trello.com/c/cyMbcLS3

There is still lots of info we can add to our OpenAPI spec to make it complete. Noting which fields are required in the requests sent to us is one.

This change is a first pass at adding this information. It was based on a call between @cruikshanks and the team's BA on 26 Nov so is limited to what could be done in the time.