Add netZeroValueInvoice field to schemas
This was added as part of https://github.com/DEFRA/charging-module-api/pull/162 which was an update to the API to prevent zero value invoices appearing in the bill run export files.

Because the code for the returning transaction data is essentially 'give me everything in the table' this field will now be included in reponses for transactions.