Start adding `required` details to request schemas (#29)
https://trello.com/c/cyMbcLS3

There is still lots of info we can add to our OpenAPI spec to make it complete. Noting which fields are required in the requests sent to us is one.

This change is a first pass at adding this information. It was based on a call between @cruikshanks and the team's BA on 26 Nov so is limited to what could be done in the time.

It also

- adds the missing `newLicence` field to post requests. This was spotted during the review with the team BA that this field was missing from the transaction post request schema.

- adds `clientId` to transaction details

Pull request https://github.com/DEFRA/charging-module-api/pull/172 added functionality for accepting client transaction IDs when a transaction POST is made. This will allow us to prevent duplicate transactions being created for a client giving them the confidence to re-try any failed requests rather than delete and then try again.

- adds netZeroValueInvoice field to schemas

This was added as part of https://github.com/DEFRA/charging-module-api/pull/162 which was an update to the API to prevent zero value invoices appearing in the bill run export files.

Because the code for the returning transaction data is essentially 'give me everything in the table' this field will now be included in responses for transactions.

- adds extra validation details for eiucSource

Spotted during the review with the team BA that we have special validation requirements around eiucSource when creating a transaction