Remove duplicate eiuc field
We added a duplicate field in https://github.com/DEFRA/sroc-service-team/pull/29 because we thought this was the only way to provide a different description in the transaction post schema to the one set in the field.

However, we have since learnt you can actually override the description in a field by just add 'description' directly into the schema (as this change does).

So this reverts the previous change to remove the duplication.