Fix Postman documentation after changes in OpenAPI (#30)
When the guide was first written we were just intending to maintain a single version of the spec for publishing to [SwaggerHub](https://swagger.io/tools/swaggerhub/). Since then we have moved to generating and retaining a version of the spec for each release of the CHA.

So this change updates to the guide to take account of this.