Add OpenAPI spec for version 2
We are currently working on [V2 of the API](https://github.com/defra/sroc-charging-module-api). As part of that the endpoints will be different. In some cases we are amending their behaviour, others are completely new!

To aid communication and feedback about the changes it would help to have an accessible version of the proposed spec. So this change updates the project to split out version 1 and 2 of the specs. We can then generate a version 2 for uploading to SwaggerHub.