Move version 1 because of issue with extension
We're heavily reliant on VSCode OpenAPI extensions when working with the spec. We use them to preview and compile the spec. It's turns out they require the spec file to be called `openapi.yml`. If we change the name they all stop working. So we've had to re-think our project structure to cater for this.