Add generate bill run summary endpoint (#35)
One of the key changes in Version 2 is that we are breaking up the functionality covered in the Version 1 view bill run endpoint.

At this time V1

- will generate the summary if one does not exist
- return a holding response until the summary is generated
- return a view of the bill run that includes _all_ transactions

On that last point, we have bill runs with more than 9,000 transactions. When you add all the invoice summary information the response becomes massive!

So the first part of breaking it up is to create a dedicated `generate-summary` endpoint.

We not actually built it yet. But for this endpoint its not a case of if, but when. So we are adding it now to the open API spec to aid with discussion about it.

* Also includes a fix - In version 2 we are being consistent with how bill runs are referenced (they are 2 words!)