Add view bill run endpoint (#36)
One of the key changes in Version 2 is that we are breaking up the functionality covered in the Version 1 view bill run endpoint.

At this time V1

- will generate the summary if one does not exist
- return a holding response until the summary is generated
- return a view of the bill run that includes _all_ transactions

On that last point, we have bill runs with more than 9,000 transactions. When you add all the invoice summary information the response becomes massive!

We've already updated the Version 2 spec with the proposed 'generate summary' endpoint. This change adds details on view bill run, and how we intend to change it from version 1.

We will now just return what bill run and invoice information we hold in the database; no holding response and no hiding the data if the summary is not generated.

The structure of the response will always be the same, the key element being `status` will always be returned.

N.B. We've not actually built it yet. But for this endpoint, it's not a case of if, but when. So we are adding it now to the open API spec to aid with discussion about these changes.