Add invoice and licence delete endpoints to spec (#38)
As part of Version 2 of the API invoices and licences will be actual records in the DB with links to the relevant bill runs and transactions. This is all part of the changes we are making to fix the performance issues in the API.

Because of this, it means we can add endpoint dedicated to there deletion that work based on ID rather than as a query param. The benefit to us is

- We can build the endpoint to be specific to the task, for example, deleting an invoice and all the work needed to update the bill run and its summary details
- We can be better assured we are deleting the right thing as the client system has provided the ID of the record, rather than just a reference

Exactly where these endpoints will sit is still unknown. For example, should they also sit under `../bill-runs/` or should we look to have a `/{regime}/invoices` path? So these changes add them under the `suggested` tag for now.

* Note - As we still need to agree on the content we have gone with just some rudimentary examples for now and held off adding the invoice and licence schemas until we have something more definitive agreed.