Add Invoice Ids and Licence details to bill run
As part of version 2 of the API we'll have actual invoice and licence records. And we intend to base responses on them, particularly the view bill run endpoint response.

We are also looking to require client systems to use the invoice and licence ID's when they wish them to be deleted. However, we've spotted we currently never provide them. That will make using them a little challenging!

So this change updates the view bill run endpoint details to include

- the id of each invoice
- an array of licences linked to the invoice

Regards the licences, as this will be a new element in the response we need to agree just how much detail is needed. So for now the details should be considered suggested rather then definitive.