Update view bill run to match invoices
Each `bill-run + customer + financial-year` combination will result in a distinct invoice. Rather than add complex logic to break this up we should amend the response to match the data we have and how its structured.

This also drops the grouping certain fields in the bill run as `summary`. It's nice for human eyes but unnecessary for the API response.