Add Invoice Ids and Licence details to bill run (#39)
As part of version 2 of the API we'll have persisted invoice and licence records. And we intend to base responses on them, particularly the view bill run endpoint response.

We are also looking to require client systems to use the invoice and licence ID's when they wish them to be deleted. However, we've spotted we currently never provide them. That will make using them a little challenging!

So this change updates the view bill run endpoint details to include

- the id of each invoice
- an array of licences linked to the invoice

Regards the licences, as this will be a new element in the response we need to agree just how much detail is needed. So for now the details should be considered suggested rather than definitive.

Each `bill-run + customer + financial-year` combination will result in a distinct invoice. Rather than add complex logic to break this up we need to amend the response to match the data we have and how it's structured.

We also drop the grouping of certain fields in the bill run under `summary`. It's nice for human eyes but unnecessary for the API response.

* Notes

- Add licence number to view licence endpoint (Bit of an oversight we're correcting here!)
- Remove v2 bill run schema - A bit like some of the other endpoints, whilst the responses are still being defined and agreed it makes sense not to carry the overhead of maintaining examples and the schema