Add bill run status endpoint to v2 spec
The workflow around bill runs is

- create bill run
- add transactions
- generate summary
- keep checking bill run status until summary is complete
- get bill run view

At the moment the only endpoint to check the status of the bill run is the view bill run. We are intending to remove the transaction detail to make it more performant. But we'll still have to collate all invoice and licence detail to provide a full response.

This seems overkill and will impact performance if all a client system actually needs to know is if the status has gone from `generating` to `summarised`. So this change adds details of a new endpoint it is likely we'll add to V2 that can be used to just check the status of a bill run.