Revert change to billrunId case
The case _should_be `billRunId`. However, after changing it Swagger flagged it as an issue because the case no longer matched the case of the param defined in the path file. This would also mean needing to update the field in the params file, and everything in `version_1`.

So, for now we are reverting this change and will tackle it separately.