Add missing properties to v2 bill run response
Thoughat this time we are not trying to define the final view bill run response (we've not built it yet!) we also know that the conversation could be diverted if people construe it in that way. So this change adds some properties to the view bill run response that exist in V1, and which we fully expect to include in V2, but which we haven't added to our examples in the spec.