Add Transaction type and reference
These are values again currently held against the transaction in V1 but actually apply at invoice level. They are not generated and applied until a bill run is sent in V1, so in our examples they would never be populated with anything. But we also intend to populate them when the `/send` request is made so it makes things clearer to include the in the details now.