Add missing properties to v2 bill run response (#41)
Though at this time we are not trying to define the final view bill run response (we've not built it yet!) we also know that the conversation could be diverted if people construe it in that way. So this change adds some properties to the view bill run response that exist in V1, and which we fully expect to include in V2, but which we haven't added to our examples in the spec.

- summarised flag

Came up in our planning today on the dev team. We are going to want to track what invoices we have summarised so _at this time_ we are intending to add a `summarised` boolean flag to do this.

It might also be useful to client systems so including it 🤷

- netZeroValue flag

This is a flag we include at the transaction level in V1, but actually is there to indicate if the invoice the transaction belongs to is a net-zero value invoice. As we have an invoice in V2 we'll track it there (which means we can also drop `invoice` from the name!)

- transaction type and reference

These are values again currently held against the transaction in V1 but actually apply at invoice level. They are not generated and applied until a bill run is sent in V1, so in our examples, they would never be populated with anything. But we also intend to populate them when the `/send` request is made so it makes things clearer to include the in the details now.