Add transaction file ref to View bill run response (#43)
`transactionFileReference` is something that gets populated when the `/send` request is made. But as part of our intent to keep the structure the same for the view bill run it will make things clearer if we include it in our examples, even though it will be empty.

So this updates the Version 2 spec to include it in the examples.