Add notes about docker hub integration
https://trello.com/c/jHORNdnQ

We have now made [version 2](https://github.com/defra/sroc-charging-module-api) available as a docker image in [Docker Hub](https://hub.docker.com/r/environmentagency/sroc-charging-module-api). As part of that we have produced a guide on how to get started, along with an example `docker-compose.yml` and `makefile`.

Normally, this information would all be added to the README on docker hub. But we have also enabled automated builds. This means you can always grab whatever is on `main` as a docker image plus whenever we add a tag docker hub automatically builds a matching image. Sweet!

The downside is the standard behaviour of an automated build is to [overwrite the README in docker hub with the one from the source repo](https://github.com/docker/hub-feedback/issues/325)! 😩🤦

So, at somepoint we'll want to update the API read me to make sense in both contexts. But for now it also means we need a place to store details about the integration plus the guide and example files we have produced.

This change covers adding all that here.