Re-structure groups
Normally we structure by domain, for example 'bill runs'. But to try and make it clearer what the status of an endpoint is as we develop V2 we are temporarily grouping things by build status.

This should make it easier to understand for both teams what the current status of an endpoint is to help planning and prioritisation.