Expand descriptions for view invoice and licence
This mirrors that last discussion with WRLS. We intend for all interactions to be ID based rather than using filters. Plus we ideally wish to structure the data we return based on how it's held in our DB.

But as to what fields are returned it makes no odds to us.