Update SROC V2 API spec docs to latest position (#45)
https://trello.com/c/jHORNdnQ

As part of making [version 2](https://github.com/DEFRA/sroc-charging-module-api) available to the WRLS team so they can start their migration, we have also committed to keeping the OpenAPI documents as up to date as possible.

We are also conscious they are a 'work in progress' constantly changing depending on what progress we have made in the API. So, this change will attempt to get us into a better position going forward. It will

- update everything that needs updating to match the current position
- try to be consistent with how we present details (to schema or not to schema, that is the question!)
- regroup things on build status rather than function (everything looks likely to end up under bill runs anyway)

Plus any other housekeeping tasks along we spot. Change after this will be smaller and specific, promise!