Correct V2 view bill run API examples (#47)
Whilst working on the latest changes we spotted one inconsistency with the current examples and what we actually return (it's `zeroLineCount` in the invoice details, not `zeroValueLineCount`).

Also, [Develop Approve Bill Run (v2)](https://trello.com/c/SmiI8YaG) is going to make some changes to the bill run structure which will affect the examples. We'll just use `status:` to show a bill run is 'approved', no need for a separate field and makes things less complex in the code.

So, we cover that change here as well and add a new example to help demonstrate it.

Finally, we spotted that the bill run status endpoint examples were still using `summarised`.