Update view bill run invoices
We update the example based on an actual response from the API. We also update the description to reflect this is now a working version, not just a standin example.

Finally, we add some relevant example error responses. There are other related to authentication, authorisation, and issues with the bill run that need adding at some point. But there is value in adding ones applicable to trying to view a bill run invoice now.