Update view bill run invoice API spec example (#48)
We are just on the cusp of making the view bill run invoice endpoint available. What we had as the example before was just a placeholder. We now have a working version so can update the spec to use what it currently returns.

We can still expect some feedback and changes once WRLS have had a chance to review it. But it at least represents the first version they will see.