Add test view transaction endpoint API details
We have recently added a 'test' specific view transaction endpoint. Like our view regime and authorised system endpoints we apply no logic to the formatting of the response from the DB. The data is returned as is.

It also includes all data held in the related bill run, invoice and licence records. The intent is to give easy access to everything related to a transaction to support testing. As we make changes to the DB record details the endpoint will continue to 'just' work.