Remove counts from V2 view bill run Open API spec (#58)
We have recently amended the fields that the view bill run endpoint returns. This updates the Open API spec to match.

We also spotted that we didn't have an example for when a bill run has been `sent` so we add that as well.

Finally, we spotted we had mistakenly overwritten the root file (openapi.yml) with the compiled output! So we fix that as well.