Update V2 Open API spec to latest
There are no real changes to any of the API end points. We have added some more examples to cover when a bill run has been determined as `billing_not_required`.

The key changes are moving some of the endpoints to the status of `tested` to reflect the work @anwarisse1 has been doing.