Add invoice rebill endpoint to V2 OpenAPI spec (#67)
Add details for the new bill run invoice rebilling endpoint to the V2 OpenAPI spec. At this time there is currently an endpoint available, but it's not doing all the checks outlined here or returning the response we've documented. This is why it is recorded as being 'suggested'. We'll bump it to available just as soon as we have hooked the endpoint up to the new services.