Update main openapi.yml
Remove all references and tags that were used when the spec was in 'build status mode'. Also remove references to old V1 (it is wiped from our memory, there is only one API!)

Also, added a section about our Docker image so we have a pointer to it from here.