Remove V1 Dockerhub docs and support
Version 1 of the API never went live and was effectively ditched in favour of V2 back in 2020.

We're a few weeks out from going live with V2 so it just doesn't make sense to keep maintaining documentation and support for something that was defunct before it ever went live.

This change removes all references to V1's dockerhub repo and the support files we provide.