Remove version 1 draft and rename v2
As version 1 is defunct it means there will never be another version of it so maintaining a draft for it is redundant. As we are now only maintaining 1 draft there is also no need to keep referring to it as the V2 draft version. It can now just be the 'draft' version!