Remove V1 Dockerhub docs and support (#71)
Version 1 of the API never went live and was effectively ditched in favour of V2 back in 2020.

We're a few weeks out from going live with V2 so it just doesn't make sense to keep maintaining documentation and support for something that was defunct before it ever went live.

This change removes all references to V1's DockerHub repo and the support files we provide.

Notes - We recently amended the npm scripts in the projects package.json. However we forgot to update the makefile example we provide. This also includes the corrections to that.