Update V2 OpenAPI spec ready for API release (#69)
We've done all the endpoints and they are now in a stable state so we are ready to fully update the Open API V2 spec. This means

- re-organising groupings to be based on features rather than implementation status
- adding full schemas to all the endpoints
- adding any outstanding endpoint changes

We also clean out the schema files to only include items referenced by the version 2 endpoints. And, as version 1 is now defunct we remove all the spec files related to maintaining it; we'll never produce another version of it so we don't need them.