Split out release process and update for V2 API
With Version 1 of the API now being properly defunct and replaced by Version 2 we need to update the release process document to stop referring to the old version.

In review, though there is a lot of similarities in the process for releasing the 2 services, trying to cover both in one document makes things hard to understand.

So, this change does 2 things

- it splits the release process in 2 so both the [SROC TCM](https://github.com/DEFRA/sroc-tcm-admin) and the [SROC Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) have their own dedicated documents
- it removes all references to the old V1 API and instead is based on V2