Add initial info on Workflows to CHA OpenAPI spec
We have descriptions and schemas for each of the endpoints but the OpenAPI spec doesn't provide any info on how they are supposed to hang together.

This change adds an initial attempt at description the workflows we expect client systems to follow and how they overlap with the Chargung Module.