Fix Open API spec customer file details (#82)
In [Add details about the customer file to API spec](https://github.com/DEFRA/sroc-service-team/pull/77) we added details on the customer file we generate and send to SSCL SOP to the Open API spec.

But immediately after we spotted some errors with the example data used. This change corrects those errors.