Add missing 4xx response to Open API spec
We have some examples of the `4xx` responses the [SROC Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) might return but its not complete.

There are some shared responses, for example, when the bill run is unknown that are not included which we feel are needed to make the spec complete.

So, this change goes through the endpoints adding any missing `4xx` responses we are aware of.