Team name change!
Having finally shipped the initial [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) which currently is only used by the WRLS team, we're thinking ahead to working with other services on things other than SROC.

With that in mind we have officially (ish) renamed ourselves.

- the service we look after shall be hence forth known as **Charging Module** (made up of the TCM, CM, and Rules Service)
- we shall be hence forth know as the **Charging Module Team**
- we shall no longer use the prefix **SROC** on everything

This change just updates the root README and any other obvious references. In the future we'll do the harder task of updating our repo names and anything we find that links to them.