Team name change! (#86)
Having finally shipped the initial [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) which currently is only used by the WRLS team, we're thinking ahead to working with other services on things other than SROC.

With that in mind, we have officially (ish) renamed ourselves.

- the service we look after shall be henceforth known as **Charging Module** (made up of the TCM, CM, and Rules Service)
- we shall be henceforth known as the **Charging Module Team**
- we shall no longer use the prefix **SROC** on everything

This change just updates the root README and any other obvious references. In the future, we'll do the harder task of updating our repo names and anything we find that links to them.

- Where we referenced the 'SROC team' before we now reference the 'Charging Module team'.
- Where we referenced the 'SROC Charging Module API' before we now reference the 'Charging Module API'
- Where we referenced the 'SROC TCM' before as now reference the 'Tactical Charging Module'

Some references have remained as they are linked to repo names which will take a bit of work to rename.