Clarified the details around minimum charge
Recent discussions between the 2 teams (CM & WRLS) have clarified the scenarios that minimum charge is applied, and it's not alway just at annual bill run.

We've updated the focus to be less about the specifics of when its applied and more on how grouping works when it is applied emphasising it's on the client system to set the flag.