Add TCM RfC example to docs
https://eaflood.atlassian.net/browse/CMEA-91

We added some guidance and an example for the CHA back in [Add section and guides for RfC's](https://github.com/DEFRA/sroc-service-team/pull/91). At the same time we added some stub pages for the RfC and the rules service.

We're now ready to raise an actual RfC for the TCM. So, it seems a great time to also knock off adding a TCM example to these guides.