Fix Customer Files endpoint ID's
We hadn't spotted that [Document list customer files endpoint](https://github.com/DEFRA/sroc-service-team/pull/102) had inadvertently duplicated the endpoint ID of the existing test endpoint.

This changes fixes that and updates `draft.yml`.

**Note**

We have a number of alignment changes in `draft.yml` as well. This was because the JSON was converted to YAML using [yq](https://mikefarah.gitbook.io/yq/) instead of (https://www.json2yaml.com/) as the site was down at the time of making the change.

We don't mind. In fact, we think it does a better job! For reference yq was installed using `brew install yq` and the command to get yq to convert our json was `yq eval -P openapi/openapi.json > openapi/versions/draft.yml`.