Update admin generate bill run OpenAPI details (#103)
https://github.com/DEFRA/sroc-charging-module-api/pull/570

We've updated the test generate bill run endpoint in the API

- a new invoice type of `standard` which just adds debit transactions had been added
- a new `transactionMultiplier` property can now be included in the request

The last is the critical change. Previously, the generator determined just how many transactions were assigned to each invoice (currently 3). Recently, we have had a need to create bill runs for test purposes that have much larger numbers of transactions. It's not just about lots of invoices (great for `/generate` testing), but invoices with lots of transactions (needed for `DELETE` testing).

So, to remain consistent we still by default add 3 transactions per invoice. But now you can multiply that by a value. For example, a `transactionMultiplier` of 3 for a `zero-value` invoice would result in 9 transactions being created and assigned to the invoice.

This change updates the OpenAPI documentation to reflect these changes.