Add new README for the CHA
We wanted to host the CHA release flow infographic we developed and this seemed the best place. We also wanted to document the commitments we have made to the WRLS team around changes to the CHA and specifically breaking changes. Again, this seems as good a place as any.

At this time, we still think a responsible API would allow some time for client systems to test and prepare their services before the API is updated. But we are concious that most folks just want us to get our changes our ASAP.

We're also a small team with other commitments (both personal and professional). So, we haven't committed to a hard timeframe in our release flow. Signing off a release in one sprint and then scheduling it in the next we think

- gives that period of time to client teams
- still gets changes out relatively quickly
- doesn't add undue stress or process to what is a small team