Add releasing document to TCM
When splitting out the existing doc we found we had some good notes about what actually needs to be done on the day of release plus what to do if something goes wrong.

Rather than loosing it we have decided to add a third document called `release.md` and add a reference to their contet in the root README.

This just covers the TCM but we are going to split up the CHA next using the same structure.