Update releasing guidance (#108)
https://eaflood.atlassian.net/browse/CMEA-133

Since going 'live' in June we have been refining the way we release the CHA and the TCM. We now treat sign off separate from submitting the RfC and scheduling the release. We have also made some tweaks to how we do both.

This change updates our releasing guidance to reflect the latest changes and bring it up to date.