Add alternate method to find RS standard change
https://eaflood.atlassian.net/browse/CMEA-97

We added an example RfC for the Rules Service in [PR 115](https://github.com/DEFRA/sroc-service-team/pull/115). This is based on us having a approved Standard Change in myIT. The standard change acts as a template from which future RfC's are generated.

Initial feedback though highlighted that for whatever reason not everyone has access to the standard change catalogue option in their menu. Fortunately, they can see the `All Templates` option which can be used to select the template.

This change documents that alternate approach in the guide.